requirements.txt
celery
django-celery-beat
redis
selenium
requests
python-dotenv
flower
fastapi
uvicorn
google-genai
openai
Pillow
pycryptodome